### 2019-12-20 #1975 @whedon
[PRE REVIEW]: partition: A fast and flexible framework for data reduction in R
**Submitting author:** @malcolmbarrett (<a href="http://orcid.org/0000-0003-0299-5825">Malcolm Barrett</a>)
**Repository:** <a href="https://github.com/USCbiostats/partition" target ="_blank">https://github.com/USCbiostats/partition</a>
**Version:** 0.1.1
