### 2019-12-20 9b1887d
use dev version of ggplot2, which addresses #3
