### 2019-12-31 #1975 @malcolmbarrett
Thanks all, and happy new year! If another reviewer is needed, I took a look at the list and here are some folks who might be good fits: dcassol, clauswilke, rcannood. Otherwise, people interested in dimension reduction, performant computing, bioinformatics, or generally working with genetic data may be interested.
