### 2020-01-03 #1975 @csoneson
👋 @dcassol, @clauswilke, @rcannood - would one of you be interested in reviewing this submission for JOSS?

> partition: A fast and flexible framework for data reduction in R
