### 2020-01-03 USCbiostats/partition#6 @clauswilke
JOSS review: Consider adding issue template
From https://github.com/openjournals/joss-reviews/issues/1991:

I suggest you add an issue template and/or a few sentences in the README describing how to file issues.
