### 2020-01-03 #1991 @clauswilke
Comments on the paper:

1. I don't understand the first sentence, "Data are increasingly wider". What does that mean? People are increasingly reluctant to store their data in long format? I don't think that's what you're trying to say, but that's what this sentence means to me. I suggest you write a first sentence that is less confusing, e.g., "Data are increasingly feature rich, ..." and then explain what "feature rich" means, i.e., many dimensions of observations per observational unit.

2. I don't know what the "Partition framework" is. I suggest you add 1-2 sentences at the end of the 1st paragraph or the beginning of the 2nd, giving a high-level description of this framework and describing how it differs, e.g., from PCA or t-SNE.

3. "While many tools exist in R for data reduction": Here it would be good to cite some packages that implement such tools.

4. I find that the first sentence of the second paragraph is difficult to read, due to the reference to R in the middle of the phrase "R package". I suggest you rephrase, e.g., "The package `partition` provides an implementation of the Partition framework for the statistical programming language R (R Core Team, 2019)."

5. You're not consistent in whether or not you place a space before references. I think there should be a space in front of every reference.
