### 2020-01-03 USCbiostats/partition#7 @clauswilke
JOSS review: Improvements to introductory vignette
From https://github.com/openjournals/joss-reviews/issues/1991:

I feel the introductory vignette could do a better job describing what partition actually does and why it is useful. The vignette starts out with a very abstract and technical description of the algorithm and then provides an example using package functions to both generate and analyze the data. Since I don't know the package functions at this point in my reading (I just started reading the introduction), and since there is no comparison with some other methods, I'm immediately lost.

A couple of specific suggestions for improvements:

1. Start with a very simple, high-level description of the method in words, so I know where things are going. (Example: "Many dimension reduction approaches, such as PCA, attempt to create a small set of new variables by combining several existing variables. By contrast, the Partition framework drops the least informative variables and retains the most informative variables as is.")

2. Maybe start with an example of a well known or highly intuitive dataset where the block correlation structure is immediately obvious and where I can interpret the results from partition. E.g., a data set of human body measurements (height, weight, leg length, arm length, etc.).

3. With whatever example dataset you use, it would be helpful to compare the output of partition to some other approach readers may be more familiar with, e.g. PCA.

4. It would be helpful to visualize the block correlation structure, e.g. by drawing a correlogram.

5. It might be helpful to add a reference to the 2019 Bioinformatics paper, so readers know where to look if they want to learn more about this method.
