### 2020-01-17 #1991 @lmweber
Thank you for your submission.

My main comment relates to possible prior publication. The authors previously published a methods paper in August 2019 (Millstein et al., 2019, [Bioinformatics](https://academic.oup.com/bioinformatics/advance-article/doi/10.1093/bioinformatics/btz661/5554652)), describing the computational methodology and introducing the `partition` R package, which was made available via [CRAN](https://cran.r-project.org/web/packages/partition/index.html) at that time.

Could the authors please provide further details about why a new submission to JOSS is now needed? E.g. have any major updates been made to the methodology? Otherwise, I believe this would constitute dual publication, and the submission should be withdrawn.

Could the editors please also provide further guidance about what constitutes dual publication in the context of a JOSS submission? (e.g. first publishing a software package in JOSS, and then using it in a biological applications paper would presumably be fine). Apologies if I am misinterpreting this issue. @csoneson

I also have a few other more specific comments:

- The package successfully installs from the provided GitHub link, and the examples in the vignette run without error. The package appears to be well structured, documented, and tested.

- Is there a reason why the JOSS submission refers only to the GitHub version of the package, and not the CRAN version? For most users, it will be easier to install from CRAN.

- As suggested by @clauswilke above (and in issue USCbiostats/partition#7), it would be useful to expand the introductory text at the start of the vignette, to more clearly spell out the need for the package, and e.g. to explain what "agglomerative partitioning" means for readers who are not familiar with this.

- Also as suggested by @clauswilke, the wording "Data are increasingly wider" is confusing. It would be clearer to refer to this as "high-dimensional" instead of "wide" data.

- The title refers to the `partition` package as "fast". However, this is not demonstrated in the text. Some additional details should be included to justify this claim.

- Is there a reason why the license details were split up into two files (`LICENSE` and `LICENSE.md`)? Usually it would be simpler to put this all within `LICENSE`.

- Regarding authorship, the previous publication (Millstein et al., 2019, [Bioinformatics](https://academic.oup.com/bioinformatics/advance-article/doi/10.1093/bioinformatics/btz661/5554652)) included a number of additional authors. Could the authors provide additional details about why these authors were not included here? (e.g. if they were not involved with the development of the software).

- As mentioned by @clauswilke (and in issue USCbiostats/partition#6), it would be useful to provide some details for users on how to submit issues.

- There is a missing quotation mark in `# install.packages("remotes)` at the start of the vignette and the GitHub readme.

I have left several tick-boxes in the reviewer checklist empty for now, for the issues described above.
