### 2020-01-17 #1991 @malcolmbarrett
Thank you for the helpful comments, @lmweber and @clauswilke. I will address them at the beginning of February. However, the question of dual publication seems urgent enough to reply immediately. So let me address that.

The Bioinformatics paper is about the framework and includes simulations and examples pertaining to it. We do mention that there *is* an R package, but that paper does *not* introduce the R package. In the JOSS paper, what we are describing is the software that was developed. Likewise, authorship pertains to the developers of the *software*, not the framework or its analysis. I think this is quite in line with the spirit of JOSS, but if this is truly an editorial concern, we are happy to discuss it further.
