### 2020-01-17 #1991 @clauswilke
My opinion is that having a publication specifically about the R package is appropriate. Different authorship is also appropriate.
