### 2020-01-18 #1991 @csoneson
> My main comment relates to possible prior publication. The authors previously published a methods paper in August 2019 (Millstein et al., 2019, Bioinformatics), describing the computational methodology and introducing the partition R package, which was made available via CRAN at that time.

Thanks for bringing this up @lmweber. I have looked at the _Bioinformatics_ paper, and I think this is acceptable for _JOSS_, given that the previous paper rather describes the algorithm/framework, while the _JOSS_ submission focuses on the software (which is also explicitly being reviewed here).
