### 2020-01-18 #1991 @lmweber
Ok, thank you for the clarification @csoneson . Seems like this is my misunderstanding then.

In this case we can ignore this part of my review, and focus on the specific comments at the end.

Apologies for the confusion @malcolmbarrett .
