### 2020-02-06 b382caa
fix partial match in `measure_min_icc()`
