### 2020-02-06 41ce69f
update news and roxygen
