+++ .github/workflows/pkgdown.yaml
@@ -1,7 +1,7 @@
 on:
   push:
     branches: master
-  pr:
+  pull_request:
     branches: master
 
 name: Pkgdown
