+++ pkgdown.yml
@@ -1,0 +1,7 @@
+pandoc: 2.7.3
+pkgdown: 1.4.1.9000
+pkgdown_sha: aceba8b49f4d20fcb9ca11549b6c477143aca5f4
+articles:
+  extending-partition: extending-partition.html
+  introduction-to-partition: introduction-to-partition.html
+
