+++ .github/workflows/pkgdown.yaml
@@ -1,8 +1,6 @@
 on:
   push:
     branches: master
-  pull_request:
-    branches: master
 
 name: Pkgdown
 
