### 2020-02-26 b07bdca
add performance section to README
