+++ README.Rmd
@@ -1,5 +1,36 @@
 ---
 output: github_document
+references:
+- id: R-partition
+  type: article-journal
+  author:
+  - family: Millstein
+    given: Joshua
+  - family: Battaglin
+    given: Francesca
+  - family: Barrett
+    given: Malcolm
+  - family: Cao
+    given: Shu
+  - family: Zhang
+    given: Wu
+  - family: Stintzing
+    given: Sebastian
+  - family: Heinemann
+    given: Volker
+  - family: Lenz
+    given: Heinz-Josef
+  issued:
+  - year: 2020
+  title: 'Partition: A surjective mapping approach for dimensionality reduction'
+  title-short: Partition
+  container-title: Bioinformatics
+  page: 676-681
+  volume: '36'
+  issue: '3'
+  URL: 10.1093/bioinformatics/btz661
+params:
+  invalidate_cache: false
 ---
 
 <!-- README.md is generated from README.Rmd. Please edit that file -->
@@ -9,7 +40,8 @@ knitr::opts_chunk$set(
   collapse = TRUE,
   comment = "#>",
   fig.path = "man/figures/README-",
-  out.width = "100%"
+  out.width = "100%",
+  dpi = 320
 )
 ```
 
@@ -40,7 +72,7 @@ remotes::install_github("USCbiostats/partition")
 
 ## Example
 
-```{r}
+```{r example}
 library(partition)
 set.seed(1234)
 df <- simulate_block_data(c(3, 4, 5), lower_corr = .4, upper_corr = .6, n = 100)
@@ -70,13 +102,102 @@ partition(df, threshold = .6, partitioner = part_icc_rowmeans)
 
 partition also supports a number of ways to visualize partitions and permutation tests; these functions all start with `plot_*()`. These functions all return ggplots and can thus be extended using ggplot2.
 
-```{r, dpi = 320}
+```{r stacked_area_chart, dpi = 320}
 plot_stacked_area_clusters(df) +
   ggplot2::theme_minimal(14)
 ```
+
+## Performance
+
+partition has been meticulously benchmarked and profiled to improve performance, and key sections are written in C++ or use C++-based packages. Using a data frame with 1 million rows on a 2017 MacBook Pro with 16 GB RAM, here's how each of the built-in partitioners perform:
+
+```{r benchmarks1, eval = FALSE}
+large_df <- simulate_block_data(c(3, 4, 5), lower_corr = .4, upper_corr = .6, n = 1e6)
+
+basic_benchmarks <- microbenchmark::microbenchmark(
+  icc = partition(large_df, .3),
+  kmeans = partition(large_df, .3, partitioner = part_kmeans()),
+  minr2 = partition(large_df, .3, partitioner = part_minr2()),
+  pc1 = partition(large_df, .3, partitioner = part_pc1()),
+  stdmi = partition(large_df, .3, partitioner = part_stdmi())
+)
+```
+
+```{r secret_benchmarks1, echo = FALSE, warning=FALSE, message=FALSE}
+library(microbenchmark)
+if (params$invalidate_cache) {
+  large_df <- simulate_block_data(c(3, 4, 5), lower_corr = .4, upper_corr = .6, n = 1e6)
+  
+  basic_benchmarks <- microbenchmark::microbenchmark(
+    icc = partition(large_df, .3),
+    kmeans = partition(large_df, .3, partitioner = part_kmeans()),
+    minr2 = partition(large_df, .3, partitioner = part_minr2()),
+    pc1 = partition(large_df, .3, partitioner = part_pc1()),
+    stdmi = partition(large_df, .3, partitioner = part_stdmi())
+  )
+  
[…]
