+++ pkgdown.yml
@@ -1,6 +1,6 @@
 pandoc: 2.7.3
 pkgdown: 1.4.1.9000
-pkgdown_sha: aceba8b49f4d20fcb9ca11549b6c477143aca5f4
+pkgdown_sha: 5e873fd9e478dddd784c31d8c8f994fd4b341c2a
 articles:
   extending-partition: extending-partition.html
   introduction-to-partition: introduction-to-partition.html
