### 2020-02-26 c27e758
clarify overall goal of Partition framework in opening paragraph
