### 2020-02-26 dc875d1
reference other DR tools in R
