### 2020-02-27 4b3021f
clean up benchmark figures
