### 2020-02-27 22a485b
add `prcomp()` to list of DR funs
