+++ paper/paper.Rmd
@@ -33,7 +33,7 @@ The default partitioner uses a correlation-based distance matrix to find the pai
 
 Many tools exist in R for data reduction, including:
 
-* In the base R stats package, `princomp()` for PCA, `kmeans()` for K-Means Clustering, and `hclust()` for Hierarchical Clustering, among others [@r-base]
+* In the base R stats package, `princomp()` and `prcomp()` for PCA, `kmeans()` for K-Means Clustering, and `hclust()` for Hierarchical Clustering, among others [@r-base]
 * cluster, an R package for grouping data using Hierarchical Clustering and other strategies [@r-cluster]
 * Rtsne and tsne, R packages for t-SNE analysis [@r-Rtsne; @r-tsne]
 * uwot, an R package for Uniform Manifold Approximation and Projection (UMAP) dimensionality reduction [@r-uwot]
