### 2020-02-28 a400d60
Clarify and expand intro vignette
