+++ vignettes/introduction-to-partition.Rmd
@@ -49,13 +49,15 @@ knitr::opts_chunk$set(
   warning = FALSE,
   message = FALSE
 )
+
+options(tibble.max_extra_cols = 10)
 ```
 
 ## Introduction to the partition package
 
-partition is a fast and flexible data reduction framework [@R-partition] that uses an approach called Direct-Measure-Reduce to create agglomerative partitions that maintain the user-specified minimum level of information. Each variable starts of as its own cluster; potential reduced variables are assessed by making sure this level of information is maintained. The reduced variables are also interpretable: original variables map to one and only one variable in the reduced data set. partition is flexible, as well: how variables are selected to reduce, how information loss is measured, and the way data is reduced can all be customized. 
+partition is a fast and flexible data reduction framework for R [@R-partition]. While there are many approaches to data reduction, such as principal components analsysis (PCA) and hierarchical clustering (both supported in base R), partition attempts to create a reduced data set that is both interpretable (each raw variable maps to one and only one reduced variable) and information-rich (reduced variables must meet an information threshold). This often results in a data set that has a mix of raw features from the original data and reduced features that combine multiple raw features.
 
-partition is particularly useful for highly correlated data, such as genetic data, where there is a lot of redundancy in the information each variable lends. `simulate_block_data()` simulates data like this: blocks of correlated data that are themselves independent of the other blocks in the data:
+partition is particularly useful for highly correlated data, such as genetic data, where there is a lot of redundancy in the information each variable lends. `simulate_block_data()` simulates data like this: blocks of correlated data that are themselves independent of the other blocks in the data.
 
 ```{r}
 library(partition)
@@ -69,28 +71,62 @@ df <- simulate_block_data(
   upper_corr = .6,
   n = 100
 )
-
-df
 ```
 
-partition takes a data frame and a threshold--the minimum amount of information a reduced variable must explain to be created--and reduces the data to as few variables as possible. The threshold prevents information loss: each reduced variable must still explain at least that much.
+In a heatmap showing the correlations between the simulated variables, blocks of correlated features are clearly visible:
 
 ```{r}
-prt <- partition(df, threshold = .5)
+ggcorrplot::ggcorrplot(corr(df))
+```
+
+Many types of data follow a pattern like this. Closely related to the block correlation structure found in genetic data is that found in microbiome data. The data set `baxter_otu` has microbiome data on 172 healthy patients. Each row represents a patient, and each column represents an Operational Taxanomic Unit (OTU). OTUs are species-like relationships between bacteria determined by analyzing their RNA. Each cell in the data represents the logged-count of the OTU found in the patient's stool sample, and there are 1,234 OTUs in all.
+
+```{r}
+baxter_otu
+```
+
+While not as obvious as simulated data, blocks of correlation also appear in these data; these bacteria tend to group together or apart in the microbiomes of participants. Here are the first 200 OTUs:
+
+```{r}
+correlation_subset <- corr(baxter_otu[, 1:200])
+ggcorrplot::ggcorrplot(correlation_subset, hc.order = TRUE) + ggplot2::theme_void()
+```
+
+# Reducing data with partition
+
+Because there are many more features (OTUs) in these data than rows (patients), it's often useful to reduce data to meaningful subsets to use in statistical modeling. The main function in partition is `partition()`. `partition()` takes a data frame and a threshold--the minimum amount of information a reduced variable must explain to be created--and reduces the data to as few variables as possible. The threshold prevents information loss: each reduced variable must still explain at least greater than or equal to the information threshold.
+
+```{r}
+prt <- partition(baxter_otu, threshold = .5)
 
 prt
 ```
 
-partition creates a `tibble` with the newly reduced variables:
[…]
