+++ vignettes/introduction-to-partition.Rmd
@@ -94,7 +94,7 @@ ggcorrplot::ggcorrplot(correlation_subset, hc.order = TRUE) + ggplot2::theme_voi
 
 # Reducing data with partition
 
-Because there are many more features (OTUs) in these data than rows (patients), it's often useful to reduce data to meaningful subsets to use in statistical modeling. The main function in partition is `partition()`. `partition()` takes a data frame and a threshold--the minimum amount of information a reduced variable must explain to be created--and reduces the data to as few variables as possible. The threshold prevents information loss: each reduced variable must still explain at least greater than or equal to the information threshold.
+Because there are many more features (OTUs) in these data than rows (patients), it's useful to reduce data to meaningful subsets to use in statistical modeling. The main function in partition is `partition()`. `partition()` takes a data frame and a threshold--the minimum amount of information a reduced variable must explain to be created--and reduces the data to as few variables as possible. The threshold prevents information loss: each reduced variable must still explain at least greater than or equal to the information threshold.
 
 ```{r}
 prt <- partition(baxter_otu, threshold = .5)
@@ -123,7 +123,7 @@ Notably, these approaches can be easily combined (see below).
 
 partition uses an approach called Direct-Measure-Reduce to create agglomerative (bottom-up) partitions that maintain the user-specified minimum level of information. Each variable starts of as its own cluster; potential reduced variables are assessed by making sure this level of information is maintained. The reduced variables are also interpretable: original variables map to one and only one variable in the reduced data set. partition is flexible, as well: how variables are selected to reduce, how information loss is measured, and the way data is reduced can all be customized. 
 
-Each variable in the original data set maps to one reduced variable; in this partition, there are four reduced variables consisting of two to three variables each, as well as five of the original variables that did not get reduced because too much information would have been lost. Here are the top 20:
+Each variable in the original data set maps to one reduced variable; in this partition, there are 63 reduced variables consisting of two to seven variables each, as well as 1,076 of the original variables that did not get reduced because too much information would have been lost. Here are the top 20 clusters, ordered by how many raw features they represent:
 
 ```{r}
 plot_ncluster(prt, show_n = 20) +
@@ -152,7 +152,7 @@ unnest_mappings(prt)
 
 ## Direct-Measure-Reduce and Partitioners
 
-Partitioners are functions that tell the partition algorithm 1) what to try to reduce 2) how to measure how much information is lost from the reduction and 3) how to reduce the data. We call this approach Direct-Measure-Reduce. In partition, functions that handle 1) are thus called directors, functions that handle 2) are called metrics, and functions that handle 3) are called reducers. partition has a number of pre-specified partitioners for agglomerative data reduction, but this approach is also quite flexible. See the [vignette on extending partition](extending-partition.html) to learn more about custom partitioners.
+Partitioners are functions that tell the partition algorithm 1) what to try to reduce 2) how to measure how much information is lost from the reduction and 3) how to reduce the data. We call this approach Direct-Measure-Reduce. In partition, functions that handle 1) are thus called directors, functions that handle 2) are called metrics, and functions that handle 3) are called reducers. partition has a number of pre-specified partitioners, but this approach is also quite flexible. See the [vignette on extending partition](extending-partition.html) to learn more about custom partitioners.
 
[…]
