### 2020-03-09 4bd54ee
bump ggplot2 min version (closes #3)
