+++ pkgdown.yml
@@ -1,6 +1,6 @@
 pandoc: 2.7.3
 pkgdown: 1.4.1.9000
-pkgdown_sha: 5e873fd9e478dddd784c31d8c8f994fd4b341c2a
+pkgdown_sha: b18aa6bc9551ca78b75167abf44f81f396c9b548
 articles:
   extending-partition: extending-partition.html
   introduction-to-partition: introduction-to-partition.html
