### 2020-03-09 #1991 @malcolmbarrett
Thanks again for your patience and to the reviewers for their very helpful feedback. We have worked a substantial amount of this feedback into our paper and introductory vignette in particular, as well as other areas of the repository, which are now all available in the master branch of the repository.

## Paper updates

### Clarity

Per the suggestions from both reviewers, we have reworked the introduction to be clearer and followed their other style suggestions. We have also added a comparison to other DR methods and mentioned several R implementations thereof. You can find the PR for the paper, including the diffs, at https://github.com/USCbiostats/partition/pull/15.

### Performance

We have made two changes to clarify the performance of partition and our efforts to measure and improve it, in response to @lmweber's comment about speed. We added a short description of our efforts (e.g. re-writing key components in C++) to the paper. Additionally, we have added several benchmarks to the `README` to help the user understand performance of the package under a few different scenarios. Please see the PR at https://github.com/USCbiostats/partition/pull/14

### Other paper updates

Per the suggestion from @lmweber, we have mentioned that the package is available on CRAN.

## Vignette updates

### Clarity

We have reworked the introduction to be a clearer and to include examples from PCA and a comparison of the output. Please see the PR at https://github.com/USCbiostats/partition/pull/17.

### Data and examples

One of our biggest changes was the addition of real-world data sets to the package, which we are excited to expand on in future releases. We included and exported several related data sets from a study on the microbiome, all starting with `baxter_*`. The most notable is `baxter_otu`, which we use in the intro vignette. Please see the PR at https://github.com/USCbiostats/partition/pull/16.

We demonstrated functionality in partition by using this data instead of simulated data, and we demonstrated how the simulated block data mimics real world data. We also included correlation heatmaps of both the microbiome data and simulated data to help the user understand how the features relate.

Our thanks to @clauswilke for making these suggestions, which we feel greatly help the vignette.

### Citation of Bioinformatics paper

We have added a citation of our Bioinformatics paper to the vignette. Additionally, we have added a citation to the `DESCRIPTION` file and to `inst/CITATION` (to render in `citation("partition")`), which we promised CRAN we would do when the paper was out, anyway. Please see the PR at https://github.com/USCbiostats/partition/pull/12

## Other changes

We have added an issue template and mentioned the contributor guidelines and CoC in the `README` (PR: https://github.com/USCbiostats/partition/pull/8)

We have corrected the `install.packages` typo pointed out by @lmweber (https://github.com/USCbiostats/partition/commit/60d2eb0862cbd5e0cf6473a35b5797e5dc15ed8b)

We have also made a few changes unrelated to the suggestions here. But FYI, since the review: 1) We now depend on the most current version of ggplot2 (we had previously depended on the GitHub version due to a bug fix required for one of our plot methods). 2) Our pkgdown site is now deployed automatically via GitHub Actions. 3) We fixed a minor partial name match for an argument in one function.

## Other comments

From @lmweber: "Is there a reason why the license details were split up into two files (`LICENSE` and `LICENSE.md`)? Usually it would be simpler to put this all within `LICENSE`."

Thank you for the attention to detail! Yes, this is because of a CRAN limitations with regards to `LICENSE`. Full license details are often put in build-ignored `LICENSE.md` files, as included in our repo, to supplement this file. From [R Packages, 2nd ed.](https://r-pkgs.org/whole-game.html):

> [The usethis package] puts a copy of the full license in LICENSE.md and adds this file to .Rbuildignore. It’s considered a best practice to include a full license in your package’s source, such as on GitHub, but CRAN disallows the inclusion of this file in a package tarball

Our thanks again to both reviewers. As expected, the feedback was extremely helpful.
