### 2020-03-16 #1991 @lmweber
The authors have comprehensively addressed all of the previous comments raised. I have updated the tick boxes in the reviewer checklist above.

A couple of additional minor comments that I noticed in the update (but which should not stop acceptance):

- PCA should be "principal component analysis" not "principal components analysis"

- the reference to the `cluster` R package (page 2 of paper) could perhaps mention that this is a "recommended" R package, instead of just a regular R package

Thanks @arfon @csoneson for the message regarding COVID-19. I hope everyone stays safe.
