### 2020-03-17 #1991 @arfon
@csoneson & @malcolmbarrett - [this PR](https://github.com/USCbiostats/partition/pull/18) fixes that DOI issue that Whedon is complaining about.
