### 2020-03-17 #1991 @csoneson
Thanks @arfon.

@malcolmbarrett - in addition to this PR, could you please take a look at the other references:
* there are a few places where you need case protection (using {}); such as R, Barnes-Hut and journal titles
* please provide a link to the Cluster package for consistency
* the Partition publication details should be updated

Next, could you please:
- [ ] create a tagged release of the software
- [ ] make a Zenodo archive. Please make sure that the title and authors of the archive are identical to those of the paper (you can also add ORCIDs in Zenodo)
- [ ] Report the archived version number and the DOI of the Zenodo archive back here
