### 2020-03-17 363852d
fix typo and add cluster detail
