+++ paper/references.bib
@@ -58,6 +58,7 @@ abstract = "Precision medicine proposes to individualize the practice of medicin
   author = {Martin Maechler and Peter Rousseeuw and Anja Struyf and Mia Hubert and Kurt Hornik},
   year = {2019},
   note = {R package version 2.1.0 --- For new features, see the 'Changelog' file (in the package source)},
+  url = {https://CRAN.R-project.org/package=cluster},
 }
 
 @Manual{r-Rtsne,
