+++ paper/references.bib
@@ -2,8 +2,8 @@
     author = {Millstein, Joshua and Battaglin, Francesca and Barrett, Malcolm and Cao, Shu and Zhang, Wu and Stintzing, Sebastian and Heinemann, Volker and Lenz, Heinz-Josef},
     title = {Partition: a surjective mapping approach for dimensionality reduction},
     journal = {Bioinformatics},
-    year = {2019},
-    month = {08},
+    year = {2020},
+    month = {02},
     abstract = "{Large amounts of information generated by genomic technologies are accompanied by statistical and computational challenges due to redundancy, badly behaved data and noise. Dimensionality reduction (DR) methods have been developed to mitigate these challenges. However, many approaches are not scalable to large dimensions or result in excessive information loss.The proposed approach partitions data into subsets of related features and summarizes each into one and only one new feature, thus defining a surjective mapping. A constraint on information loss determines the size of the reduced dataset. Simulation studies demonstrate that when multiple related features are associated with a response, this approach can substantially increase the number of true associations detected as compared to principal components analysis, non-negative matrix factorization or no DR. This increase in true discoveries is explained both by a reduced multiple-testing challenge and a reduction in extraneous noise. In an application to real data collected from metastatic colorectal cancer tumors, more associations between gene expression features and progression free survival and response to treatment were detected in the reduced than in the full untransformed dataset.Freely available R package from CRAN, https://cran.r-project.org/package=partition.Supplementary data are available at Bioinformatics online.}",
     issn = {1367-4803},
     doi = {10.1093/bioinformatics/btz661},
@@ -30,7 +30,7 @@
 	abstract = {Advances in omics technologies - such as genomics, transcriptomics, proteomics and metabolomics - have begun to enable personalized medicine at an extraordinarily detailed molecular level. Individually, these technologies have contributed medical advances that have begun to enter clinical practice. However, each technology individually cannot capture the entire biological complexity of most human diseases. Integration of multiple technologies has emerged as an approach to provide a more comprehensive view of biology and disease. In this Review, we discuss the potential for combining diverse types of data and the utility of this approach in human health and disease. We provide examples of data integration to understand, diagnose and inform treatment of diseases, including rare and common diseases as well as cancer and transplant biology. Finally, we discuss technical and other challenges to clinical implementation of integrative omics.},
 	language = {eng},
 	number = {5},
-	journal = {Nature reviews. Genetics},
+	journal = {Nature Reviews. Genetics},
 	author = {Karczewski, Konrad J and Snyder, Michael P},
 	month = may,
 	year = {2018},
@@ -62,7 +62,7 @@ abstract = "Precision medicine proposes to individualize the practice of medicin
 }
 
 @Manual{r-Rtsne,
-  title = {{Rtsne}: T-Distributed Stochastic Neighbor Embedding using Barnes-Hut Implementation},
+  title = {{Rtsne}: {T-Distributed} Stochastic Neighbor Embedding using {Barnes-Hut} Implementation},
   author = {Jesse H. Krijthe},
   year = {2015},
   note = {R package version 0.15},
@@ -70,7 +70,7 @@ abstract = "Precision medicine proposes to individualize the practice of medicin
 }
 
 @Manual{r-tsne,
-  title = {tsne: T-Distributed Stochastic Neighbor Embedding for R (t-SNE)},
+  title = {tsne: T-Distributed Stochastic Neighbor Embedding for {R} ({t-SNE})},
   author = {Justin Donaldson},
   year = {2016},
   note = {R package version 0.1-3},
@@ -78,7 +78,7 @@ abstract = "Precision medicine proposes to individualize the practice of medicin
 }
[…]
