### 2020-03-17 #1991 @malcolmbarrett
Ok, the first part is done, as are the fixes recommended by @lmweber.

I also created a release and the Zenodo archive, for which the doi is 10.5281/zenodo.3713898. I'm not sure what you mean by archived version number. Do you mean the software version? We archived a development version, `0.1.1.9000`, but the current CRAN release is `0.1.1`.
