+++ README.Rmd
@@ -49,6 +49,9 @@ knitr::opts_chunk$set(
 [![Travis build status](https://travis-ci.org/USCbiostats/partition.svg?branch=master)](https://travis-ci.org/USCbiostats/partition) [![Build status](https://ci.appveyor.com/api/projects/status/fofb9um8jqyyi6av?svg=true)](https://ci.appveyor.com/project/malcolmbarrett/partition-qaqc6)
  [![Coverage status](https://codecov.io/gh/USCbiostats/partition/branch/master/graph/badge.svg)](https://codecov.io/github/USCbiostats/partition?branch=master)
  [![CRAN status](https://www.r-pkg.org/badges/version-ago/partition)](https://cran.r-project.org/package=partition)
+ [![status](https://joss.theoj.org/papers/6d4569d6cea0a16c87703134fc78fe59/status.svg)](https://joss.theoj.org/papers/6d4569d6cea0a16c87703134fc78fe59)
+ [![DOI](https://zenodo.org/badge/178615892.svg)](https://zenodo.org/badge/latestdoi/178615892)
+
  <!-- badges: end -->
  
 # partition
