+++ pkgdown.yml
@@ -4,5 +4,5 @@ pkgdown_sha: 630bf9f05b53a6b63d658bf2643855a8b7eda0c0
 articles:
   extending-partition: extending-partition.html
   introduction-to-partition: introduction-to-partition.html
-last_built: 2020-03-17T21:20Z
+last_built: 2020-03-17T21:27Z
 
