### 2020-03-18 #1991 @csoneson
@malcolmbarrett - apologies for the unclarity on my side. The archived version number I meant is the release tag. In your case it's `joss` - we typically prefer semantic versioning (so the tag would be e.g. `v0.1.1.9000`), but if you have a reason for preferring `joss`, that would be ok too.

For the Zenodo record, please fix the title and author list so that they are identical to those of the paper.
