### 2020-03-18 #1991 @malcolmbarrett
Oh, sorry, it looks like my changes weren't published. The title and authors should match now. I'm ok with using `v0.1.1.9000`; I only meant to distinguish the release from a proper CRAN release. I changed it to the development version number for both the GitHub release and the Zenodo release. Thanks for clarifying; that was helpful!
