### 2020-03-18 a2cbe5a
add bioinformatics paper details; consistent page numbers
