+++ paper/references.bib
@@ -4,6 +4,9 @@
     journal = {Bioinformatics},
     year = {2020},
     month = {02},
+	  volume = {36},
+	  number = {3},
+	  pages = "676-681",
     abstract = "{Large amounts of information generated by genomic technologies are accompanied by statistical and computational challenges due to redundancy, badly behaved data and noise. Dimensionality reduction (DR) methods have been developed to mitigate these challenges. However, many approaches are not scalable to large dimensions or result in excessive information loss.The proposed approach partitions data into subsets of related features and summarizes each into one and only one new feature, thus defining a surjective mapping. A constraint on information loss determines the size of the reduced dataset. Simulation studies demonstrate that when multiple related features are associated with a response, this approach can substantially increase the number of true associations detected as compared to principal components analysis, non-negative matrix factorization or no DR. This increase in true discoveries is explained both by a reduced multiple-testing challenge and a reduction in extraneous noise. In an application to real data collected from metastatic colorectal cancer tumors, more associations between gene expression features and progression free survival and response to treatment were detected in the reduced than in the full untransformed dataset.Freely available R package from CRAN, https://cran.r-project.org/package=partition.Supplementary data are available at Bioinformatics online.}",
     issn = {1367-4803},
     doi = {10.1093/bioinformatics/btz661},
@@ -35,14 +38,14 @@
 	month = may,
 	year = {2018},
 	keywords = {Humans, Metabolomics/*methods, Precision Medicine/*methods, Proteomics/*methods},
-	pages = {299--310}
+	pages = {299-310}
 }
 
 @article{malod-dogin-2018,
 title = "Precision medicine ― A promising, yet challenging road lies ahead",
 journal = "Current Opinion in Systems Biology",
 volume = "7",
-pages = "1 - 7",
+pages = "1-7",
 year = "2018",
 note = "• Future of systems biology• Genomics and epigenomics",
 issn = "2452-3100",
